+++ noesis/examples/direction_demo/__init__.py
@@ -1,3 +1,0 @@
-"""Direction layer demo package."""
-
-from .policy import GuardrailsPolicy  # noqa: F401